services/kettle/requirements.txt
rsa
uvloop
aiohttp
gunicorn
pyyaml